+++ packages/flutter_android_lifecycle/ios/Classes/FlutterAndroidLifecyclePlugin.m
@@ -1,20 +1,9 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 #import "FlutterAndroidLifecyclePlugin.h"
 
 @implementation FlutterAndroidLifecyclePlugin
-+ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  FlutterMethodChannel* channel = [FlutterMethodChannel
-      methodChannelWithName:@"flutter_android_lifecycle"
-            binaryMessenger:[registrar messenger]];
-  FlutterAndroidLifecyclePlugin* instance = [[FlutterAndroidLifecyclePlugin alloc] init];
-  [registrar addMethodCallDelegate:instance channel:channel];
-}
-
-- (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
-  if ([@"getPlatformVersion" isEqualToString:call.method]) {
-    result([@"iOS " stringByAppendingString:[[UIDevice currentDevice] systemVersion]]);
-  } else {
-    result(FlutterMethodNotImplemented);
-  }
-}
-
++ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {}
 @end